pyproject.toml
[tool.poetry]
name = "TessaractOCR"
version = "0.1.0"
description = "Sample project TessaractOCR"
authors = ["Tiểu mễ <anh_pc@live.com>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
tesserocr = "^2.5.1"
Pillow = "^7.1.2"
pdf2image = "^1.13.1"
yapf = "^0.30.0"
isort = "^4.3.21"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"